Organized account-related classes into account package.
