+++ core/src/main/java/com/aarves/bluepages/adapter/presenters/AccountPresenter.java
@@ -1,8 +1,8 @@
 package com.aarves.bluepages.adapter.presenters;
 
-import com.aarves.bluepages.usecase.interactors.AccountOutputBoundary;
-import com.aarves.bluepages.usecase.interactors.LoginResult;
-import com.aarves.bluepages.usecase.interactors.RegisterResult;
+import com.aarves.bluepages.usecase.interactors.account.AccountOutputBoundary;
+import com.aarves.bluepages.usecase.interactors.account.LoginResult;
+import com.aarves.bluepages.usecase.interactors.account.RegisterResult;
 
 public class AccountPresenter implements AccountOutputBoundary {
     private AccountView accountView;
