+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/account/AccountInputBoundary.java
@@ -1,4 +1,4 @@
-package com.aarves.bluepages.usecase.interactors;
+package com.aarves.bluepages.usecase.interactors.account;
 
 public interface AccountInputBoundary {
     boolean isLoggedIn();
