+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/account/RegisterResult.java
@@ -1,4 +1,4 @@
-package com.aarves.bluepages.usecase.interactors;
+package com.aarves.bluepages.usecase.interactors.account;
 
 public enum RegisterResult {
     SUCCESS,
